Fixed some bugs for the docker golang builder

+++ golang/onbuild-1.14.0/entrypoint.sh
@@ -19,25 +19,31 @@ validate() {
     && die "container should not run as root"
   [[ $(id -u) -eq $EXPECTED_NON_ROOT_UID ]] \
       || die "user execing the container is not the non root user expected"
-  . /.bash_profile && command -v gox \
+  . /.bash_profile >/dev/null && command -v gox \
       || die "gox not in the path"
-  . /.bash_profile && command -v dep \
+  . /.bash_profile >/dev/null && command -v dep \
       || die "dep not in the path"
   ok
 }
 
 build_and_package() {
   echo "Starting to build binaries"
-  . /.bash_profile \
-    && CGO_ENABLED=0 gox -osarch='linux/amd64 linux/386 darwin/amd64 darwin/386' -rebuild -tags='netgo' -ldflags='-w -extldflags "-static"'
+  . /.bash_profile >/dev/null \
+    && CGO_ENABLED=0 gox -osarch='linux/amd64 darwin/amd64' -rebuild -tags='netgo' -ldflags='-w -extldflags "-static"'
+  . /.bash_profile >/dev/null \
+    && CGO_ENABLED=0 gox -osarch='linux/386 darwin/386' -rebuild -tags='netgo' -ldflags='-w -extldflags "-static"'
   echo "Starting to build package"
+  [[ -f ${PROJECT}_linux_386 ]] || die "${PROJECT}_linux_386 binary not present"
   mv ${PROJECT}_linux_386 $ARTIFACT_VOLUME_DIR
+  [[ -f ${PROJECT}_linux_amd64 ]] || die "${PROJECT}_linux_amd64 binary not present"
   mv ${PROJECT}_linux_amd64 $ARTIFACT_VOLUME_DIR
+  [[ -f ${PROJECT}_darwin_386 ]] || die "${PROJECT}_darwin_386 binary not present"
   mv ${PROJECT}_darwin_386 $ARTIFACT_VOLUME_DIR
+  [[ -f ${PROJECT}_darwin_amd64 ]] || die "${PROJECT}_darwin_amd64 binary not present"
   mv ${PROJECT}_darwin_amd64 $ARTIFACT_VOLUME_DIR
   cd $ARTIFACT_VOLUME_DIR
   tar czf $PROJECT.tar.gz *
-  rm ${PROJECT}_linux_386 $PROJECT_darwin_386 ${PROJECT}_linux_amd64 $PROJECT_darwin_amd64
+  rm ${PROJECT}_linux_386 ${PROJECT}_darwin_386 ${PROJECT}_linux_amd64 ${PROJECT}_darwin_amd64
   echo "Done building the package"
   ok
 }
